+++ weblate/checks/same.py
@@ -143,6 +143,7 @@ class SameCheck(TargetCheck):
     def should_ignore(self, source, unit):
         """Check whether given unit should be ignored."""
         from weblate.checks.flags import TYPED_FLAGS
+        from weblate.glossary.models import get_glossary_terms
 
         if "strict-same" in unit.all_flags:
             return False
@@ -166,8 +167,21 @@ class SameCheck(TargetCheck):
             or "©" in source
         ):
             return True
+
+        # Strip glossary terms
+        stripped = source
+        if "check-glossary" in unit.all_flags:
+            # Extract untranslatable terms
+            terms = [
+                re.escape(term.source)
+                for term in get_glossary_terms(unit)
+                if "read-only" in term.all_flags
+            ]
+            if terms:
+                stripped = re.sub("|".join(terms), "", source, flags=re.IGNORECASE)
+
         # Strip format strings
-        stripped = strip_string(source, unit.all_flags)
+        stripped = strip_string(stripped, unit.all_flags)
 
         # Strip placeholder strings
         if "placeholders" in TYPED_FLAGS and "placeholders" in unit.all_flags:
@@ -183,6 +197,7 @@ class SameCheck(TargetCheck):
         for word in SPLIT_RE.split(stripped.lower()):
             if not test_word(word, extra_ignore):
                 return False
+
         return True
 
     def should_skip(self, unit):