+++ docs/changes.rst
@@ -13,6 +13,7 @@ Not yet released.
 * Rewritten metrics storage.
 * Added support for searching strings by position.
 * Documentation improvements.
+* Unchanged translation check can honor untranslatable terms from glossary.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/93?closed=1>`__.